add-3-tests-and-separate-test-file
