+++ contracts/nois-oracle/src/contract.rs
@@ -449,731 +449,3 @@ fn execute_add_round(
 fn calculate_bot_incentive_coefficient() -> u128 {
     1 //For now we just incentivise with the base/minimum price. We need to implement here the incentive logic and who gets how much
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use cosmwasm_std::testing::{
-        mock_dependencies, mock_env, mock_ibc_channel_close_init, mock_ibc_channel_connect_ack,
-        mock_ibc_channel_open_init, mock_ibc_channel_open_try, mock_info, MockApi, MockQuerier,
-        MockStorage,
-    };
-    use cosmwasm_std::{coin, from_binary, OwnedDeps};
-    use nois_protocol::{APP_ORDER, BAD_APP_ORDER};
-
-    const CREATOR: &str = "creator";
-
-    fn setup() -> OwnedDeps<MockStorage, MockApi, MockQuerier> {
-        let mut deps = mock_dependencies();
-        let msg = InstantiateMsg {
-            test_mode: true,
-            bot_incentive_base_price: 1_000_000u128.into(),
-            native_denom: "unois".to_string(),
-        };
-        let info = mock_info(CREATOR, &[]);
-        let res = instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-        assert_eq!(0, res.messages.len());
-        deps
-    }
-
-    // connect will run through the entire handshake to set up a proper connect and
-    // save the account (tested in detail in `proper_handshake_flow`)
-    fn connect(mut deps: DepsMut, channel_id: &str, account: impl Into<String>) {
-        let _account: String = account.into();
-
-        let handshake_open = mock_ibc_channel_open_init(channel_id, APP_ORDER, IBC_APP_VERSION);
-        // first we try to open with a valid handshake
-        ibc_channel_open(deps.branch(), mock_env(), handshake_open).unwrap();
-
-        // then we connect (with counter-party version set)
-        let handshake_connect =
-            mock_ibc_channel_connect_ack(channel_id, APP_ORDER, IBC_APP_VERSION);
-        let res = ibc_channel_connect(deps.branch(), mock_env(), handshake_connect).unwrap();
-        assert_eq!(res.messages.len(), 0);
-        assert_eq!(res.events.len(), 1);
-        assert_eq!(
-            res.events[0],
-            Event::new("ibc").add_attribute("channel", "connect"),
-        );
-    }
-
-    //
-    // Instantiate tests
-    //
-
-    #[test]
-    fn instantiate_works() {
-        let mut deps = mock_dependencies();
-
-        let msg = InstantiateMsg {
-            test_mode: true,
-            bot_incentive_base_price: 1_000_000u128.into(),
-            native_denom: "unois".to_string(),
-        };
-        let info = mock_info("creator", &[]);
-        let res = instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-        assert_eq!(0, res.messages.len());
-    }
-
-    //
-    // Execute tests
-    //
-    fn register_bot(deps: DepsMut, info: MessageInfo) {
-        let register_bot_msg = ExecuteMsg::RegisterBot {
-            moniker: "Best Bot".to_string(),
-        };
-        execute(deps, mock_env(), info, register_bot_msg).unwrap();
-    }
-
-    #[test]
-    fn add_round_verifies_and_stores_randomness() {
-        let mut deps = mock_dependencies();
-
-        let info = mock_info("creator", &[]);
-        let msg = InstantiateMsg {
-            test_mode: true,
-            bot_incentive_base_price: 1_000_000u128.into(),
-            native_denom: "unois".to_string(),
-        };
-        instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-        let info = mock_info("anyone", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-
-        let msg = ExecuteMsg::AddRound {
-            // curl -sS https://drand.cloudflare.com/public/72785
-            round: 72785,
-            previous_signature: HexBinary::from_hex("a609e19a03c2fcc559e8dae14900aaefe517cb55c840f6e69bc8e4f66c8d18e8a609685d9917efbfb0c37f058c2de88f13d297c7e19e0ab24813079efe57a182554ff054c7638153f9b26a60e7111f71a0ff63d9571704905d3ca6df0b031747").unwrap(),
-            signature: HexBinary::from_hex("82f5d3d2de4db19d40a6980e8aa37842a0e55d1df06bd68bddc8d60002e8e959eb9cfa368b3c1b77d18f02a54fe047b80f0989315f83b12a74fd8679c4f12aae86eaf6ab5690b34f1fddd50ee3cc6f6cdf59e95526d5a5d82aaa84fa6f181e42").unwrap(),
-        };
-        execute(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-        let response: BeaconResponse = from_binary(
-            &query(deps.as_ref(), mock_env(), QueryMsg::Beacon { round: 72785 }).unwrap(),
-        )
-        .unwrap();
-        assert_eq!(
-            response.beacon.unwrap().randomness.to_hex(),
-            "8b676484b5fb1f37f9ec5c413d7d29883504e5b669f604a1ce68b3388e9ae3d9"
-        );
-    }
-
-    #[test]
-    fn unregistered_bot_can_add_round() {
-        let mut deps = mock_dependencies();
-
-        let info = mock_info("creator", &[]);
-        let msg = InstantiateMsg {
-            test_mode: true,
-            bot_incentive_base_price: 1_000_000u128.into(),
-            native_denom: "unois".to_string(),
-        };
-        instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-        let msg = ExecuteMsg::AddRound {
-                // curl -sS https://drand.cloudflare.com/public/72785
-                round: 72785,
-                previous_signature: HexBinary::from_hex("a609e19a03c2fcc559e8dae14900aaefe517cb55c840f6e69bc8e4f66c8d18e8a609685d9917efbfb0c37f058c2de88f13d297c7e19e0ab24813079efe57a182554ff054c7638153f9b26a60e7111f71a0ff63d9571704905d3ca6df0b031747").unwrap(),
-                signature: HexBinary::from_hex("82f5d3d2de4db19d40a6980e8aa37842a0e55d1df06bd68bddc8d60002e8e959eb9cfa368b3c1b77d18f02a54fe047b80f0989315f83b12a74fd8679c4f12aae86eaf6ab5690b34f1fddd50ee3cc6f6cdf59e95526d5a5d82aaa84fa6f181e42").unwrap(),
-            };
-        let info = mock_info("unregistered_bot", &[]);
-        let response = execute(deps.as_mut(), mock_env(), info, msg.clone()).unwrap();
-        let randomness_attr = response
-            .attributes
-            .iter()
-            .find(|Attribute { key, .. }| key == "randomness")
-            .unwrap();
-        assert_eq!(
-            randomness_attr.value,
-            "8b676484b5fb1f37f9ec5c413d7d29883504e5b669f604a1ce68b3388e9ae3d9"
-        );
-    }
-    #[test]
-    fn add_round_fails_for_broken_signature() {
-        let mut deps = mock_dependencies();
-
-        let info = mock_info("creator", &[]);
-        let msg = InstantiateMsg {
-            test_mode: true,
-            bot_incentive_base_price: 1_000_000u128.into(),
-            native_denom: "unois".to_string(),
-        };
-        instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-        let info = mock_info("anyone", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        let msg = ExecuteMsg::AddRound {
-            // curl -sS https://drand.cloudflare.com/public/72785
-            round: 72785,
-            previous_signature: hex::decode("a609e19a03c2fcc559e8dae14900aaefe517cb55c840f6e69bc8e4f66c8d18e8a609685d9917efbfb0c37f058c2de88f13d297c7e19e0ab24813079efe57a182554ff054c7638153f9b26a60e7111f71a0ff63d9571704905d3ca6df0b031747").unwrap().into(),
-            signature: hex::decode("3cc6f6cdf59e95526d5a5d82aaa84fa6f181e4").unwrap().into(), // broken signature
-        };
-        let result = execute(deps.as_mut(), mock_env(), info, msg);
-        match result.unwrap_err() {
-            ContractError::InvalidSignature {} => {}
-            err => panic!("Unexpected error: {:?}", err),
-        };
-    }
-
-    #[test]
-    fn add_round_fails_for_invalid_signature() {
-        let mut deps = mock_dependencies();
-
-        let info = mock_info("creator", &[]);
-        let msg = InstantiateMsg {
-            test_mode: true,
-            bot_incentive_base_price: 1_000_000u128.into(),
-            native_denom: "unois".to_string(),
-        };
-        instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-        let info = mock_info("anyone", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        let msg = ExecuteMsg::AddRound {
-            // curl -sS https://drand.cloudflare.com/public/72785
-            round: 1111, // wrong round
-            previous_signature: hex::decode("a609e19a03c2fcc559e8dae14900aaefe517cb55c840f6e69bc8e4f66c8d18e8a609685d9917efbfb0c37f058c2de88f13d297c7e19e0ab24813079efe57a182554ff054c7638153f9b26a60e7111f71a0ff63d9571704905d3ca6df0b031747").unwrap().into(),
-            signature: hex::decode("82f5d3d2de4db19d40a6980e8aa37842a0e55d1df06bd68bddc8d60002e8e959eb9cfa368b3c1b77d18f02a54fe047b80f0989315f83b12a74fd8679c4f12aae86eaf6ab5690b34f1fddd50ee3cc6f6cdf59e95526d5a5d82aaa84fa6f181e42").unwrap().into(),
-        };
-        let result = execute(deps.as_mut(), mock_env(), info, msg);
-        match result.unwrap_err() {
-            ContractError::InvalidSignature {} => {}
-            err => panic!("Unexpected error: {:?}", err),
-        };
-    }
-
-    #[test]
-    fn add_round_succeeds_multiple_times() {
-        let mut deps = mock_dependencies();
-
-        let info = mock_info("creator", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        let msg = InstantiateMsg {
-            test_mode: true,
-            bot_incentive_base_price: 1_000_000u128.into(),
-            native_denom: "unois".to_string(),
-        };
-        instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-        let msg = ExecuteMsg::AddRound {
-            // curl -sS https://drand.cloudflare.com/public/72785
-            round: 72785,
-            previous_signature: HexBinary::from_hex("a609e19a03c2fcc559e8dae14900aaefe517cb55c840f6e69bc8e4f66c8d18e8a609685d9917efbfb0c37f058c2de88f13d297c7e19e0ab24813079efe57a182554ff054c7638153f9b26a60e7111f71a0ff63d9571704905d3ca6df0b031747").unwrap(),
-            signature: HexBinary::from_hex("82f5d3d2de4db19d40a6980e8aa37842a0e55d1df06bd68bddc8d60002e8e959eb9cfa368b3c1b77d18f02a54fe047b80f0989315f83b12a74fd8679c4f12aae86eaf6ab5690b34f1fddd50ee3cc6f6cdf59e95526d5a5d82aaa84fa6f181e42").unwrap(),
-        };
-
-        // Execute 1
-        let info = mock_info("anyone", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        let response = execute(deps.as_mut(), mock_env(), info, msg.clone()).unwrap();
-        let randomness_attr = response
-            .attributes
-            .iter()
-            .find(|Attribute { key, .. }| key == "randomness")
-            .unwrap();
-        assert_eq!(
-            randomness_attr.value,
-            "8b676484b5fb1f37f9ec5c413d7d29883504e5b669f604a1ce68b3388e9ae3d9"
-        );
-
-        // Execute 2
-        let info = mock_info("someone else", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        let response = execute(deps.as_mut(), mock_env(), info, msg).unwrap();
-        let randomness_attr = response
-            .attributes
-            .iter()
-            .find(|Attribute { key, .. }| key == "randomness")
-            .unwrap();
-        assert_eq!(
-            randomness_attr.value,
-            "8b676484b5fb1f37f9ec5c413d7d29883504e5b669f604a1ce68b3388e9ae3d9"
-        );
-    }
-
-    #[test]
-    fn add_round_fails_when_same_bot_submits_multiple_times() {
-        let mut deps = mock_dependencies();
-
-        let info = mock_info("creator", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        let msg = InstantiateMsg {
-            test_mode: true,
-            bot_incentive_base_price: 1_000_000u128.into(),
-            native_denom: "unois".to_string(),
-        };
-        instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-        let msg = ExecuteMsg::AddRound {
-            // curl -sS https://drand.cloudflare.com/public/72785
-            round: 72785,
-            previous_signature: HexBinary::from_hex("a609e19a03c2fcc559e8dae14900aaefe517cb55c840f6e69bc8e4f66c8d18e8a609685d9917efbfb0c37f058c2de88f13d297c7e19e0ab24813079efe57a182554ff054c7638153f9b26a60e7111f71a0ff63d9571704905d3ca6df0b031747").unwrap(),
-            signature: HexBinary::from_hex("82f5d3d2de4db19d40a6980e8aa37842a0e55d1df06bd68bddc8d60002e8e959eb9cfa368b3c1b77d18f02a54fe047b80f0989315f83b12a74fd8679c4f12aae86eaf6ab5690b34f1fddd50ee3cc6f6cdf59e95526d5a5d82aaa84fa6f181e42").unwrap(),
-        };
-
-        // Execute A1
-        let info = mock_info("bot_alice", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        execute(deps.as_mut(), mock_env(), info, msg.clone()).unwrap();
-        // Execute B1
-        let info = mock_info("bot_bob", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        execute(deps.as_mut(), mock_env(), info, msg.clone()).unwrap();
-
-        // Execute A2
-        let info = mock_info("bot_alice", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        let err = execute(deps.as_mut(), mock_env(), info, msg.clone()).unwrap_err();
-        assert!(matches!(err, ContractError::SubmissionExists));
-        // Execute B2
-        let info = mock_info("bot_alice", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        let err = execute(deps.as_mut(), mock_env(), info, msg).unwrap_err();
-        assert!(matches!(err, ContractError::SubmissionExists));
-    }
-
-    #[test]
-    fn query_beacons_asc_works() {
-        let mut deps = mock_dependencies();
-
-        let info = mock_info("creator", &[]);
-        let msg = InstantiateMsg {
-            test_mode: true,
-            bot_incentive_base_price: 1_000_000u128.into(),
-            native_denom: "unois".to_string(),
-        };
-        instantiate(deps.as_mut(), mock_env(), info.to_owned(), msg).unwrap();
-
-        let info = mock_info("anyone", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        add_test_rounds(deps.as_mut(), "anyone");
-
-        // Unlimited
-        let BeaconsResponse { beacons } = from_binary(
-            &query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::BeaconsAsc {
-                    start_after: None,
-                    limit: None,
-                },
-            )
-            .unwrap(),
-        )
-        .unwrap();
-        let response_rounds = beacons.iter().map(|b| b.round).collect::<Vec<u64>>();
-        assert_eq!(response_rounds, [72785, 72786, 72787]);
-
-        // Limit 2
-        let BeaconsResponse { beacons } = from_binary(
-            &query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::BeaconsAsc {
-                    start_after: None,
-                    limit: Some(2),
-                },
-            )
-            .unwrap(),
-        )
-        .unwrap();
-        let response_rounds = beacons.iter().map(|b| b.round).collect::<Vec<u64>>();
-        assert_eq!(response_rounds, [72785, 72786]);
-
-        // After 0
-        let BeaconsResponse { beacons } = from_binary(
-            &query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::BeaconsAsc {
-                    start_after: Some(0),
-                    limit: None,
-                },
-            )
-            .unwrap(),
-        )
-        .unwrap();
-        let response_rounds = beacons.iter().map(|b| b.round).collect::<Vec<u64>>();
-        assert_eq!(response_rounds, [72785, 72786, 72787]);
-
-        // After 72785
-        let BeaconsResponse { beacons } = from_binary(
-            &query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::BeaconsAsc {
-                    start_after: Some(72785),
-                    limit: None,
-                },
-            )
-            .unwrap(),
-        )
-        .unwrap();
-        let response_rounds = beacons.iter().map(|b| b.round).collect::<Vec<u64>>();
-        assert_eq!(response_rounds, [72786, 72787]);
-
-        // After 72787
-        let BeaconsResponse { beacons } = from_binary(
-            &query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::BeaconsAsc {
-                    start_after: Some(72787),
-                    limit: None,
-                },
-            )
-            .unwrap(),
-        )
-        .unwrap();
-        let response_rounds = beacons.iter().map(|b| b.round).collect::<Vec<u64>>();
-        assert_eq!(response_rounds, Vec::<u64>::new());
-    }
-
-    #[test]
-    fn query_beacons_desc_works() {
-        let mut deps = mock_dependencies();
-
-        let info = mock_info("creator", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        let msg = InstantiateMsg {
-            test_mode: true,
-            bot_incentive_base_price: 1_000_000u128.into(),
-            native_denom: "unois".to_string(),
-        };
-        instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-        let info = mock_info("anyone", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        add_test_rounds(deps.as_mut(), "anyone");
-
-        // Unlimited
-        let BeaconsResponse { beacons } = from_binary(
-            &query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::BeaconsDesc {
-                    start_after: None,
-                    limit: None,
-                },
-            )
-            .unwrap(),
-        )
-        .unwrap();
-        let response_rounds = beacons.iter().map(|b| b.round).collect::<Vec<u64>>();
-        assert_eq!(response_rounds, [72787, 72786, 72785]);
-
-        // Limit 2
-        let BeaconsResponse { beacons } = from_binary(
-            &query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::BeaconsDesc {
-                    start_after: None,
-                    limit: Some(2),
-                },
-            )
-            .unwrap(),
-        )
-        .unwrap();
-        let response_rounds = beacons.iter().map(|b| b.round).collect::<Vec<u64>>();
-        assert_eq!(response_rounds, [72787, 72786]);
-
-        // After 99999
-        let BeaconsResponse { beacons } = from_binary(
-            &query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::BeaconsDesc {
-                    start_after: Some(99999),
-                    limit: None,
-                },
-            )
-            .unwrap(),
-        )
-        .unwrap();
-        let response_rounds = beacons.iter().map(|b| b.round).collect::<Vec<u64>>();
-        assert_eq!(response_rounds, [72787, 72786, 72785]);
-
-        // After 72787
-        let BeaconsResponse { beacons } = from_binary(
-            &query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::BeaconsDesc {
-                    start_after: Some(72787),
-                    limit: None,
-                },
-            )
-            .unwrap(),
-        )
-        .unwrap();
-        let response_rounds = beacons.iter().map(|b| b.round).collect::<Vec<u64>>();
-        assert_eq!(response_rounds, [72786, 72785]);
-
-        // After 72785
-        let BeaconsResponse { beacons } = from_binary(
-            &query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::BeaconsDesc {
-                    start_after: Some(72785),
-                    limit: None,
-                },
-            )
-            .unwrap(),
-        )
-        .unwrap();
-        let response_rounds = beacons.iter().map(|b| b.round).collect::<Vec<u64>>();
-        assert_eq!(response_rounds, Vec::<u64>::new());
-    }
-
-    //TODO Add a test to check that an unregistered bot does not get incentives
-    //TODO Add a test to check that an registered bot gets incentives
-    //TODO Add a test to check that when a contract does not have enough balance no incentives are given
-
-    #[test]
-    fn query_submissions_works() {
-        let mut deps = mock_dependencies();
-
-        let info = mock_info("creator", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        let msg = InstantiateMsg {
-            test_mode: true,
-            bot_incentive_base_price: 1_000_000u128.into(),
-            native_denom: "unois".to_string(),
-        };
-        instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-        let info = mock_info("bot1", &[]);
-        register_bot(deps.as_mut(), info.to_owned());
-        add_test_rounds(deps.as_mut(), "bot1");
-
-        // No submissions
-        let response: SubmissionsResponse = from_binary(
-            &query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::Submissions { round: 72777 },
-            )
-            .unwrap(),
-        )
-        .unwrap();
-        assert_eq!(response.round, 72777);
-        assert_eq!(response.submissions, Vec::<_>::new());
-
-        // One submission
-        let response: SubmissionsResponse = from_binary(
-            &query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::Submissions { round: 72785 },
-            )
-            .unwrap(),
-        )
-        .unwrap();
-        assert_eq!(response.round, 72785);
-        assert_eq!(
-            response.submissions,
-            [Submission {
-                bot: Addr::unchecked("bot1"),
-                time: Timestamp::from_nanos(1571797419879305533),
-            }]
-        );
-
-        add_test_rounds(deps.as_mut(), "bot2");
-
-        // Two submissions
-        let response: SubmissionsResponse = from_binary(
-            &query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::Submissions { round: 72785 },
-            )
-            .unwrap(),
-        )
-        .unwrap();
-        assert_eq!(response.round, 72785);
-        assert_eq!(
-            response.submissions,
-            [
-                Submission {
-                    bot: Addr::unchecked("bot1"),
-                    time: Timestamp::from_nanos(1571797419879305533),
-                },
-                Submission {
-                    bot: Addr::unchecked("bot2"),
-                    time: Timestamp::from_nanos(1571797419879305533),
-                }
-            ]
-        );
-    }
-
-    /// Adds round 72785, 72786, 72787
-    fn add_test_rounds(mut deps: DepsMut, bot_addr: &str) {
-        let msg = ExecuteMsg::AddRound {
-            // curl -sS https://drand.cloudflare.com/public/72785
-            round: 72785,
-            previous_signature: HexBinary::from_hex("a609e19a03c2fcc559e8dae14900aaefe517cb55c840f6e69bc8e4f66c8d18e8a609685d9917efbfb0c37f058c2de88f13d297c7e19e0ab24813079efe57a182554ff054c7638153f9b26a60e7111f71a0ff63d9571704905d3ca6df0b031747").unwrap(),
-            signature: HexBinary::from_hex("82f5d3d2de4db19d40a6980e8aa37842a0e55d1df06bd68bddc8d60002e8e959eb9cfa368b3c1b77d18f02a54fe047b80f0989315f83b12a74fd8679c4f12aae86eaf6ab5690b34f1fddd50ee3cc6f6cdf59e95526d5a5d82aaa84fa6f181e42").unwrap(),
-        };
-        let info = mock_info(bot_addr, &[]);
-        register_bot(deps.branch(), info.to_owned());
-        execute(deps.branch(), mock_env(), info, msg).unwrap();
-        let msg = ExecuteMsg::AddRound {
-            // curl -sS https://drand.cloudflare.com/public/72786
-            round: 72786,
-            previous_signature: HexBinary::from_hex("82f5d3d2de4db19d40a6980e8aa37842a0e55d1df06bd68bddc8d60002e8e959eb9cfa368b3c1b77d18f02a54fe047b80f0989315f83b12a74fd8679c4f12aae86eaf6ab5690b34f1fddd50ee3cc6f6cdf59e95526d5a5d82aaa84fa6f181e42").unwrap(),
-            signature: HexBinary::from_hex("85d64193239c6a2805b5953521c1e7c412d13f8b29df2dfc796b7dc8e1fd795b764362e49302956a350f9385f68b68d8085fda08c2bd0528984a413db52860b408c72d1210609de3a342259d4c08f86ee729a2dbeb140908270849fd7d0dec40").unwrap(),
-        };
-        let info = mock_info(bot_addr, &[]);
-        execute(deps.branch(), mock_env(), info, msg).unwrap();
-        let msg = ExecuteMsg::AddRound {
-            // curl -sS https://drand.cloudflare.com/public/72787
-            round: 72787,
-            previous_signature: HexBinary::from_hex("85d64193239c6a2805b5953521c1e7c412d13f8b29df2dfc796b7dc8e1fd795b764362e49302956a350f9385f68b68d8085fda08c2bd0528984a413db52860b408c72d1210609de3a342259d4c08f86ee729a2dbeb140908270849fd7d0dec40").unwrap(),
-            signature: HexBinary::from_hex("8ceee95d523f54a752807f4705ce0f89e69911dd3dce330a337b9409905a881a2f879d48fce499bfeeb3b12e7f83ab7d09b42f31fa729af4c19adfe150075b2f3fe99c8fbcd7b0b5f0bb91ac8ad8715bfe52e3fb12314fddb76d4e42461f6ea4").unwrap(),
-        };
-        let info = mock_info(bot_addr, &[]);
-        execute(deps.branch(), mock_env(), info, msg).unwrap();
-    }
-
-    //
-    // IBC tests
-    //
-
-    #[test]
-    fn enforce_version_in_handshake() {
-        let mut deps = setup();
-
-        let wrong_order = mock_ibc_channel_open_try("channel-12", BAD_APP_ORDER, IBC_APP_VERSION);
-        ibc_channel_open(deps.as_mut(), mock_env(), wrong_order).unwrap_err();
-
-        let wrong_version = mock_ibc_channel_open_try("channel-12", APP_ORDER, "another version");
-        ibc_channel_open(deps.as_mut(), mock_env(), wrong_version).unwrap_err();
-
-        let valid_handshake = mock_ibc_channel_open_try("channel-12", APP_ORDER, IBC_APP_VERSION);
-        ibc_channel_open(deps.as_mut(), mock_env(), valid_handshake).unwrap();
-    }
-
-    #[test]
-    fn proper_handshake_flow() {
-        let mut deps = setup();
-        let channel_id = "channel-1234";
-
-        // first we try to open with a valid handshake
-        let handshake_open = mock_ibc_channel_open_init(channel_id, APP_ORDER, IBC_APP_VERSION);
-        ibc_channel_open(deps.as_mut(), mock_env(), handshake_open).unwrap();
-
-        // then we connect (with counter-party version set)
-        let handshake_connect =
-            mock_ibc_channel_connect_ack(channel_id, APP_ORDER, IBC_APP_VERSION);
-        let _res = ibc_channel_connect(deps.as_mut(), mock_env(), handshake_connect).unwrap();
-    }
-
-    #[test]
-    fn check_close_channel() {
-        let mut deps = setup();
-
-        let channel_id = "channel-123";
-        let account = "acct-123";
-
-        // register the channel
-        connect(deps.as_mut(), channel_id, account);
-        // assign it some funds
-        let funds = vec![coin(123456, "uatom"), coin(7654321, "tgrd")];
-        deps.querier.update_balance(account, funds);
-
-        // close the channel
-        let channel = mock_ibc_channel_close_init(channel_id, APP_ORDER, IBC_APP_VERSION);
-        let _res = ibc_channel_close(deps.as_mut(), mock_env(), channel).unwrap();
-    }
-
-    //
-    // Other
-    //
-
-    #[test]
-    fn next_round_works_for_test_mode() {
-        let mut deps = mock_dependencies();
-        let (round, source_id) = next_round(&mut deps.storage, NextRoundMode::Test).unwrap();
-        assert_eq!(round, 2183660);
-        assert_eq!(source_id, "test-mode:2183660");
-        let (round, source_id) = next_round(&mut deps.storage, NextRoundMode::Test).unwrap();
-        assert_eq!(round, 2183661);
-        assert_eq!(source_id, "test-mode:2183661");
-        let (round, source_id) = next_round(&mut deps.storage, NextRoundMode::Test).unwrap();
-        assert_eq!(round, 2183662);
-        assert_eq!(source_id, "test-mode:2183662");
-    }
-
-    #[test]
-    fn next_round_works_for_time_mode() {
-        let mut deps = mock_dependencies();
-
-        // UNIX epoch
-        let (round, _) = next_round(
-            &mut deps.storage,
-            NextRoundMode::Time {
-                base: Timestamp::from_seconds(0),
-            },
-        )
-        .unwrap();
-        assert_eq!(round, 1);
-
-        // Before Drand genesis (https://api3.drand.sh/info)
-        let (round, _) = next_round(
-            &mut deps.storage,
-            NextRoundMode::Time {
-                base: Timestamp::from_seconds(1595431050).minus_nanos(1),
-            },
-        )
-        .unwrap();
-        assert_eq!(round, 1);
-
-        // At Drand genesis (https://api3.drand.sh/info)
-        let (round, _) = next_round(
-            &mut deps.storage,
-            NextRoundMode::Time {
-                base: Timestamp::from_seconds(1595431050),
-            },
-        )
-        .unwrap();
-        assert_eq!(round, 2);
-
-        // After Drand genesis (https://api3.drand.sh/info)
-        let (round, _) = next_round(
-            &mut deps.storage,
-            NextRoundMode::Time {
-                base: Timestamp::from_seconds(1595431050).plus_nanos(1),
-            },
-        )
-        .unwrap();
-        assert_eq!(round, 2);
-
-        // Drand genesis +29s/30s/31s
-        let (round, _) = next_round(
-            &mut deps.storage,
-            NextRoundMode::Time {
-                base: Timestamp::from_seconds(1595431050).plus_seconds(29),
-            },
-        )
-        .unwrap();
-        assert_eq!(round, 2);
-        let (round, _) = next_round(
-            &mut deps.storage,
-            NextRoundMode::Time {
-                base: Timestamp::from_seconds(1595431050).plus_seconds(30),
-            },
-        )
-        .unwrap();
-        assert_eq!(round, 3);
-        let (round, _) = next_round(
-            &mut deps.storage,
-            NextRoundMode::Time {
-                base: Timestamp::from_seconds(1595431050).plus_seconds(31),
-            },
-        )
-        .unwrap();
-        assert_eq!(round, 3);
-    }
-}
